moved connections component into Footer component, as it will be repeated on new page
